+++ app/src/main/java/ua/com/radiokot/photoprism/features/gallery/search/view/model/SearchBookmarkDialogViewModel.kt
@@ -56,6 +56,8 @@ class SearchBookmarkDialogViewModel(
             isSaveButtonEnabled.value = canSave
         }
 
+        isInitialized = true
+
         log.debug {
             "initOnce(): initialized:" +
                     "\nstate=${stateSubject.value}"