+++ app/src/main/java/ua/com/radiokot/photoprism/features/gallery/logic/MediaPreviewUrlFactory.kt
@@ -1,17 +1,19 @@
 package ua.com.radiokot.photoprism.features.gallery.logic
 
+import ua.com.radiokot.photoprism.api.photos.model.PhotoPrismMergedPhoto
+
 interface MediaPreviewUrlFactory {
     fun getSmallThumbnailUrl(hash: String): String
     fun getMediumThumbnailUrl(hash: String): String
 
-    fun getPreview720Url(hash: String): String
-    fun getPreview1280Url(hash: String): String
-    fun getPreview1920Url(hash: String): String
-    fun getPreview2048Url(hash: String): String
-    fun getPreview2560Url(hash: String): String
-    fun getPreview3840Url(hash: String): String
-    fun getPreview4096Url(hash: String): String
-    fun getPreview7680Url(hash: String): String
+    fun getImagePreview720Url(hash: String): String
+    fun getImagePreview1280Url(hash: String): String
+    fun getImagePreview1920Url(hash: String): String
+    fun getImagePreview2048Url(hash: String): String
+    fun getImagePreview2560Url(hash: String): String
+    fun getImagePreview3840Url(hash: String): String
+    fun getImagePreview4096Url(hash: String): String
+    fun getImagePreview7680Url(hash: String): String
 
-    fun getMp4PreviewUrl(hash: String): String
+    fun getVideoPreviewUrl(mergedPhoto: PhotoPrismMergedPhoto): String
 }