+++ app/src/main/java/ua/com/radiokot/photoprism/features/gallery/logic/PhotoPrismPreviewUrlFactory.kt
@@ -1,8 +1,11 @@
 package ua.com.radiokot.photoprism.features.gallery.logic
 
+import ua.com.radiokot.photoprism.api.photos.model.PhotoPrismMergedPhoto
+
 class PhotoPrismPreviewUrlFactory(
     apiUrl: String,
     private val previewToken: String,
+    private val videoFormatSupport: VideoFormatSupport,
 ) : MediaPreviewUrlFactory {
     private val previewUrlBase = "${apiUrl}v1"
 
@@ -12,33 +15,63 @@ class PhotoPrismPreviewUrlFactory(
     override fun getMediumThumbnailUrl(hash: String): String =
         "$previewUrlBase/t/$hash/$previewToken/tile_500"
 
-    override fun getPreview720Url(hash: String): String =
+    override fun getImagePreview720Url(hash: String): String =
         getFitPreviewUrl(hash, 720)
 
-    override fun getPreview1280Url(hash: String): String =
+    override fun getImagePreview1280Url(hash: String): String =
         getFitPreviewUrl(hash, 1280)
 
-    override fun getPreview1920Url(hash: String): String =
+    override fun getImagePreview1920Url(hash: String): String =
         getFitPreviewUrl(hash, 1920)
 
-    override fun getPreview2048Url(hash: String): String =
+    override fun getImagePreview2048Url(hash: String): String =
         getFitPreviewUrl(hash, 2048)
 
-    override fun getPreview2560Url(hash: String): String =
+    override fun getImagePreview2560Url(hash: String): String =
         getFitPreviewUrl(hash, 2560)
 
-    override fun getPreview3840Url(hash: String): String =
+    override fun getImagePreview3840Url(hash: String): String =
         getFitPreviewUrl(hash, 3840)
 
-    override fun getPreview4096Url(hash: String): String =
+    override fun getImagePreview4096Url(hash: String): String =
         getFitPreviewUrl(hash, 4096)
 
-    override fun getPreview7680Url(hash: String): String =
+    override fun getImagePreview7680Url(hash: String): String =
         getFitPreviewUrl(hash, 7680)
 
     private fun getFitPreviewUrl(hash: String, size: Int) =
         "$previewUrlBase/t/$hash/$previewToken/fit_$size"
 
-    override fun getMp4PreviewUrl(hash: String): String =
-        "$previewUrlBase/videos/$hash/$previewToken/avc"
+    override fun getVideoPreviewUrl(mergedPhoto: PhotoPrismMergedPhoto): String {
+        // https://github.com/photoprism/photoprism/blob/2f9792e5411f6bb47a84b638dfc42d51b7790853/frontend/src/model/photo.js#L489
+
+        val file = mergedPhoto.videoFile
+        if (file?.width == null || file.height == null) {
+            return "$previewUrlBase/videos/${mergedPhoto.hash}/$previewToken/avc"
+        }
+
+        val fileCodec = file.codec ?: ""
+        val previewFormat = when {
+            (fileCodec == "hvc1" || fileCodec == "hev1")
+                    && videoFormatSupport.canPlayHevc(file.width, file.height) ->
+                "hevc"
+
+            fileCodec == "vp8" && videoFormatSupport.canPlayVp8(file.width, file.height) ->
+                "vp8"
+
+            fileCodec == "vp9" && videoFormatSupport.canPlayVp9(file.width, file.height) ->
+                "vp9"
+
+            (fileCodec == "av01" || fileCodec == "av1c")
+                    && videoFormatSupport.canPlayAv1(file.width, file.height) ->
+                "av01"
+
+            // WebM and OGV seems not supported.
+
+            else ->
+                "avc"
+        }
+
+        return "$previewUrlBase/videos/${file.hash}/$previewToken/$previewFormat"
+    }
 }